Getting everything close to finalized.
Added in the sensor board as well as fasteners for both boards and the
baseplate.

Also added in models for connectors and such to make sure our tolerances
look good.  That includes extruding a little cut off for the sensor
board pigtail because it was going to be a tight fit.

We're using #4 screws for both boards and #6 screws for the base plate.