Add in volume knob lighting passages.
Added in openings within the volume knob support walls, and in the
support wall itself, to let light out from within.  We'll have LEDs
aimed upwards toward the knob, surrounding the volume potentiometer,
which will shine up, out, and down in the space between the base and the
volume knob, hopefully providing a pleasing white light to show the unit
is powered up, and where the volume knob itself is.