Switch to the legit RCA jack model and add volume knob weight.
Finally got the STEP model for the PJRAN2X1U so that's in there now.
Also added in 7 1/2" x 1/2" set screws to the volume knob, on the
inside, to add some more heft.

Also tweaked the material of some parts to try and get a more accurate
mass of the assembly.